requirements.txt
anthropic>=0.40.0
edge-tts>=6.1.0
moviepy>=2.2.0
Pillow>=10.0.0
imageio-ffmpeg>=0.5.0
